Elevate a fifth; catch the separated neg form too
One elevated, make compare clean, address confirmed in the overlay ELF.

  OvlFunc_907_2008d10  02008d10

THE SUBTRACTION HAS TO BE WIDE.  `(unsigned short)(gState.f1c2 - 3) <= 1`
reads naturally and is wrong: with the member an `unsigned short`, gcc keeps
the whole thing in sixteen bits and emits `ldr r2, =0xfffd / add r3, r2` -- a
pooled -3 ADDED -- where the ROM has a plain `sub r3, #3`.  Reading the member
into a named `int` first widens the subtraction and the `sub` comes back.  An
explicit `(int)` cast does NOT do it; the local does.

PICK_CANDIDATES' NEW FILTER HAD A HOLE AND IT COST A SCREEN IMMEDIATELY.  Its
first version required the `mov` and its shift or negate to be adjacent, but
gcc emits all of a call's `mov`s and then all its `neg`s, so

    mov r0,#1 / mov r1,#1 / mov r2,#1 / mov r3,#0 / neg r1,r1 / neg r2,r2 / neg r0,r0

-- which is -1 built three times -- slipped through and OvlFunc_881_20097fc was
offered anyway.  The filter now matches each `neg rN, rN` back to the
`mov rN, #imm` that last set that register and reports it as [repeats -1].

Parked that one and OvlFunc_959_2009ab0, which is the second member of the
pool-load-scheduled-too-early class after OvlFunc_883_20091d8 and sits at 34 of
41 with the same shape.

+++ scratch/c8d10.c
@@ -23,7 +23,7 @@ void OvlFunc_907_2008d10(void)
 
     *(int *)(*(char **)iwram_3001ebc + (0xe0 << 1)) = 0x204;
     OvlFunc_907_2008fa0();
-    if ((unsigned short)((int)gState.f1c2 - 3) <= 1) {
+    if ((unsigned short)(gState.f1c2 - 3) <= 1) {
         if (__GetFlag(0x109) == 0) {
             a = __MapActor_GetActor(0);
             __CutsceneStart();
